Reorganize tools/mandarin
