+++ tools/mandarin/__main__.py
@@ -1,0 +1,13 @@
+#!/usr/bin/python
+# -*- coding: utf-8 -*-
+
+"""This is a script for generating mandarin.xml from CC-CEDICT.
+
+It downloads the most recent version from https://cc-cedict.org and transforms
+it into the XML format expected by sinoparserd.
+"""
+
+from __init__ import print_xml
+
+if __name__ == '__main__':
+    print_xml()
